moved get step count to util
