+++ util/util.go
@@ -1,8 +1,15 @@
 package util
 
 import (
+	"os"
+	"strconv"
+
 	"github.com/mohae/deepcopy"
 	"github.com/project-flogo/core/data/coerce"
+)
+
+const (
+	FlogoStepCount = "FLOGO_STEP_COUNT"
 )
 
 func DeepCopy(data interface{}) interface{} {
@@ -13,4 +20,17 @@ func DeepCopyMap(data map[string]interface{}) map[string]interface{} {
 	copiedData := deepcopy.Copy(data)
 	copiedMap, _ := coerce.ToObject(copiedData)
 	return copiedMap
+}
+
+// GetMaxStepCount returns the step limit
+func GetMaxStepCount() int {
+	maxStepCount := 10000000
+	envStepCount, exists := os.LookupEnv(FlogoStepCount)
+	if exists {
+		i, err := strconv.Atoi(envStepCount)
+		if err == nil {
+			return i
+		}
+	}
+	return maxStepCount
 }
